refactor(app): extract AccountFooter from DocumentSidebar
The sidebar's name implies "document list," but it had grown
account-management concerns: the post-OAuth-link URL flash parsing,
the settings-modal lifecycle, the login/logout/settings buttons,
and the modal-vs-user gating. Stripped these into a new
`AccountFooter` component sitting alongside `NetworkStatus` and
`ColorSchemeSwitcher` in the sidebar's footer.

Sidebar's responsibilities are now strictly document-list shaped:
the doc list, group headers, the convert-to-synced aria-live
region, the collapse toggle. Account state stays out — it only
reads `user` to decide whether to expose the per-row convert
affordance.

`AccountFooter` owns:

- Login / logout / open-settings buttons (rendered conditionally on
  auth state).
- The `linkFlash` + `settingsOpen` state, both lazily initialized
  from the URL so an OAuth-link redirect lands the modal open with
  the flash visible.
- The `replaceState` URL cleanup that strips the `?account_link=…`
  params after they're consumed.
- The settings-modal mount/unmount, including the "user becomes
  null → unmount" gate.

CSS: `authButton` styling moved to a new `AccountFooter.module.css`
(component now owns its own button look). Sidebar's CSS module
loses the corresponding rules.

No behavior change.
